Fix type mismatch in test assertions and concurrent test race conditions
- TestConcurrentCapReads now uses atomic counter for goroutine assertions instead of calling s.Assert() from goroutines (prevents testify suite data race)
- Add sync/atomic import to ringbuf_test.go
- Cast 0 to int64 in TestStatsAccuracy for GreaterOrEqual assertions (testify requires both args to be same type)
- All tests now pass with -race detector

+++ synctools/bounded_chan_test.go
@@ -281,8 +281,8 @@ func (s *boundedChanTestSuite) TestStatsAccuracy() {
 
 	snap := stats()
 	// Items may have been partially drained, so just verify ranges
-	s.GreaterOrEqual(snap.BufferedItems, 0)
-	s.GreaterOrEqual(snap.BufferedBytes, 0)
+	s.GreaterOrEqual(snap.BufferedItems, int64(0))
+	s.GreaterOrEqual(snap.BufferedBytes, int64(0))
 	s.Assert().Equal(int64(5), snap.MaxItems)
 	s.Assert().Equal(int64(1000), snap.MaxBytes)
